refactor(tests): consolidate duplicate fixtures and rename unused parameter
- Move api_client_no_auth_override fixture to tests/integration/conftest.py
- Remove duplicate fixture definitions from test_status_endpoint.py and test_cli_status.py
- Add test_token_service alias fixture to support consolidated api_client_no_auth_override
- Rename admin parameter to _admin in status.py (follows Python convention for dependency injection side-effects)
- Remove unused imports (Path, pytest) from test files after fixture consolidation

Addresses #299
